scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
python-dotenv==1.0.0
lxml==4.9.3